home.hcp.spec.tsx fully rewritten
