+++ packages/yourloops/lib/team/team.hook.tsx
@@ -54,6 +54,7 @@ import SpinningLoader from '../../components/loaders/spinning-loader'
 const ReactTeamContext = createContext<TeamContext>({} as TeamContext)
 
 export const PRIVATE_TEAM_ID = 'private'
+export const PRIVATE_TEAM_NAME = 'private'
 
 function TeamContextImpl(): TeamContext {
   const authHook = useAuth()
