+++ packages/yourloops/test/integration/assert/header.ts
@@ -31,7 +31,7 @@ import { type Team } from '../../../lib/team'
 import userEvent from '@testing-library/user-event'
 
 interface TeamMenuInfo {
-  selectedTeamName: string
+  selectedTeamName?: string
   isSelectedTeamPrivate?: boolean
   availableTeams: Team[]
 }
@@ -76,7 +76,10 @@ const checkUserMenu = async (header: BoundFunctions<typeof queries>, userName: s
   expect(screen.queryByTestId('user-menu')).not.toBeInTheDocument()
 }
 
-const checkTeamScopeMenu = async (header: BoundFunctions<typeof queries>, selectedTeamParams: { teamName: string, isPrivate?: boolean }, availableTeams: Team[]) => {
+const checkTeamScopeMenu = async (header: BoundFunctions<typeof queries>, selectedTeamParams: {
+  teamName: string
+  isPrivate?: boolean
+}, availableTeams: Team[]) => {
   const teamScopeMenuIcon = selectedTeamParams.isPrivate ? 'private-practice-icon' : 'medical-team-icon'
   const teamScopeMenuText = selectedTeamParams.isPrivate ? 'My private practice' : selectedTeamParams.teamName
 
@@ -126,7 +129,10 @@ const checkTeamScopeMenu2 = async (header: BoundFunctions<typeof queries>, teamM
   expect(screen.queryByTestId('team-scope-menu')).not.toBeInTheDocument()
 }
 
-export const checkHcpHeader = async (fullName: string, selectedTeamParams: { teamName: string, isPrivate?: boolean }, availableTeams: Team[]) => {
+export const checkHcpHeader = async (fullName: string, selectedTeamParams: {
+  teamName: string
+  isPrivate?: boolean
+}, availableTeams: Team[]) => {
   const header = within(screen.getByTestId('app-main-header'))
 
   expect(header.getByText('Patients')).toBeVisible()
@@ -170,4 +176,12 @@ export const checkPatientHeader = async (fullName: string) => {
 
   await checkUserMenu(header, fullName, UserRole.Patient)
   checkHeader(header)
+}
+
+export const changeTeamScope = async (currentTeamName: string, wantedTeamName: string) => {
+  const header = within(screen.getByTestId('app-main-header'))
+  await userEvent.click(header.getByText(currentTeamName))
+
+  const teamScopeMenu = within(screen.getByTestId('team-scope-menu'))
+  await userEvent.click(teamScopeMenu.getByText(wantedTeamName))
 }
