+++ packages/yourloops/test/integration/mock/team.api.mock.ts
@@ -38,6 +38,7 @@ import { type Profile } from '../../../lib/auth/models/profile.model'
 import { DEFAULT_THRESHOLDS_IN_MGDL } from '../../../components/monitoring-alert/monitoring-alert.default'
 import { type ITeam } from '../../../lib/team/models/i-team.model'
 import { Unit } from 'medical-domain'
+import { PRIVATE_TEAM_ID, PRIVATE_TEAM_NAME } from '../../../lib/team/team.hook'
 
 const {
   minVeryLowBg,
@@ -73,8 +74,6 @@ export const myFirstTeamName = 'MyFirstTeam'
 export const mySecondTeamName = 'MySecondTeam'
 export const myThirdTeamName = 'A - MyThirdTeam - to be deleted'
 export const filtersTeamName = 'Team used to test the patients filters'
-
-const privateTeamId = 'private'
 
 export const buildTeamOne = (): Team => {
   return {
@@ -212,8 +211,8 @@ export const buildFiltersTeam = (): Team => {
 
 export const buildPrivateTeam = (): Team => {
   return {
-    name: 'private',
-    id: privateTeamId,
+    name: PRIVATE_TEAM_NAME,
+    id: PRIVATE_TEAM_ID,
     code: 'private',
     type: TeamType.private,
     phone: '',
