+++ packages/yourloops/test/integration/pages/home-page/home.caregiver.spec.tsx
@@ -42,8 +42,8 @@ import DirectShareApi from '../../../../lib/share/direct-share.api'
 import { UserRole } from '../../../../lib/auth/models/enums/user-role.enum'
 import { mockUserApi } from '../../mock/user.api.mock'
 import { mockPatientApiForCaregivers } from '../../mock/patient.api.mock'
-import { checkPatientListHeader } from '../../assert/patient-list'
 import PatientApi from '../../../../lib/patient/patient.api'
+import { testPatientListHeaderCaregiver } from '../../assert/patient-list'
 
 describe('Caregiver home page', () => {
   const firstName = 'Eric'
@@ -65,7 +65,7 @@ describe('Caregiver home page', () => {
     })
     expect(await screen.findByTestId('app-main-header')).toBeVisible()
     await checkCaregiverLayout(`${firstName} ${lastName}`)
-    checkPatientListHeader(UserRole.Caregiver)
+    testPatientListHeaderCaregiver()
   })
 
   it('should filter patients correctly depending on the search value', async () => {
@@ -133,7 +133,7 @@ describe('Caregiver home page', () => {
     })
 
     await checkCaregiverLayout(`${firstName} ${lastName}`)
-    checkPatientListHeader(UserRole.Caregiver)
+    testPatientListHeaderCaregiver()
 
     const patientTableBody = screen.getByTestId('patient-list-grid')
     expect(within(patientTableBody).getAllByRole('row')).toHaveLength(5)
@@ -212,7 +212,7 @@ describe('Caregiver home page', () => {
 
     const columnSettingsButton = screen.getByTestId('column-settings-button')
 
-    checkPatientListHeader(UserRole.Caregiver)
+    testPatientListHeaderCaregiver()
     expect(screen.getByRole('columnheader', { name: 'Patient' })).toBeVisible()
     expect(screen.getByRole('columnheader', { name: 'System' })).toBeVisible()
     expect(screen.getByRole('columnheader', { name: 'Last data update' })).toBeVisible()
