+++ packages/yourloops/test/integration/use-cases/patients-vizualization-management.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2023, Diabeloop
+ * Copyright (c) 2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,20 +25,26 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-const commonJestConfig = require('../common-jest.config')
-module.exports = {
-  ...commonJestConfig,
+import {
+  type Router,
+  testPatientColumnsFiltersContent,
+  testPatientColumnsFiltersHideColumns,
+  testPatientColumnsFiltersShowColumns,
+  testPatientListContentHcp,
+  testPatientListCurrentTab,
+  testPatientListFilters,
+  testPatientListHeaderHcp,
+  testPatientListPendingTab
+} from '../assert/patient-list'
 
-  bail: true,
-
-  displayName: 'yourloops integration',
-
-  maxWorkers: '30%',
-
-  // The glob patterns Jest uses to detect test files
-  testMatch: [
-    '<rootDir>/**/*.spec.tsx'
-  ],
-
-  testTimeout: 100000
+export const testPatientsListForHcp = async (router: Router) => {
+  testPatientListHeaderHcp()
+  testPatientListContentHcp()
+  await testPatientListFilters()
+  await testPatientColumnsFiltersContent()
+  await testPatientColumnsFiltersHideColumns()
+  await testPatientColumnsFiltersShowColumns()
+  await testPatientListPendingTab(router)
+  // Must be kept at the end as this function redirects to the patient dashboard
+  await testPatientListCurrentTab(router)
 }
