+++ packages/yourloops/test/integration/use-cases/teams-management.ts
@@ -31,10 +31,15 @@ import {
   checkJoinTeamDialogDisplayErrorMessage,
   checkJoinTeamDialogPrivacyCancel
 } from '../assert/join-team'
+import { testCreateCareTeamDialog } from '../assert/team-dialog'
 
 export const testJoinTeamDialog = async () => {
   await checkJoinTeamDialog()
   await checkJoinTeamDialogCancel()
   await checkJoinTeamDialogPrivacyCancel()
   await checkJoinTeamDialogDisplayErrorMessage()
+}
+
+export const testTeamCreation = async () => {
+  await testCreateCareTeamDialog()
 }
